Generator rebuild v2
